fixed an unsigned long long -> size_t

+++ siemens_mriclient/main.cpp
@@ -788,7 +788,7 @@ int ACE_TMAIN(int argc, ACE_TCHAR *argv[] )
 
         traj = boost::shared_ptr< hoNDArray<floatd2> >(new hoNDArray<floatd2>);
 
-        std::vector<unsigned long long> trajectory_dimensions;
+        std::vector<size_t> trajectory_dimensions;
         trajectory_dimensions.push_back(ngrad);
         trajectory_dimensions.push_back(interleaves);
